+++ src/src_platform_swing/fmg/swing/draw/img/MosaicSkillOrGroupView.java
@@ -45,10 +45,10 @@ abstract class MosaicSkillOrGroupView<TImage, TImageModel extends AnimatedImageM
       }
 
       g.setComposite(AlphaComposite.SrcOver);
-      int bw = m.getBorderWidth();
+      double bw = m.getBorderWidth();
       boolean needDrawPerimeterBorder = (!m.getBorderColor().isTransparent() && (bw > 0));
       java.awt.Color borderColor = !needDrawPerimeterBorder ? null : Cast.toColor(m.getBorderColor());
-      BasicStroke bs = !needDrawPerimeterBorder ? null : new BasicStroke(bw);
+      BasicStroke bs = !needDrawPerimeterBorder ? null : new BasicStroke((float)bw);
       Stream<Pair<Color, Stream<PointDouble>>> stars = getCoords();
       stars.forEach(pair -> {
          Polygon poly = Cast.toPolygon(pair.second.collect(Collectors.toList()));
